Decouple hover logic from Main.kt
